requirements.txt
cffi==1.9.1
coreapi==2.1.1
cryptography==1.6
Django==1.8.11
django-filter==1.0.1
djangorestframework==3.5.3
djangorestframework-yaml==1.0.3
django-rest-swagger==0.3.5
django-sslserver==0.19
enum34==1.1.6
idna==2.1
ipaddress==1.0.17
itypes==1.1.0
MySQL-python==1.2.5
openapi-codec==1.2.0
pyasn1==0.1.9
pycparser==2.17
pyOpenSSL==16.2.0
PyYAML==3.12
requests==2.12.3
simplejson==3.10.0
six==1.10.0
Unidecode==0.4.19
uritemplate==3.0.0
